python_deprecated/requirements.txt
aiohttp
pydantic
Pillow
websocket-client
